590 refactor special uses e2e * move files from root into special uses
