+++ frontend/e2e/authenticated/special-uses/app.e2e-spec.ts
@@ -1,4 +1,4 @@
-import { FrontendPage, HelpMePick } from '../app.po';
+import { FrontendPage, HelpMePick } from './app.po';
 import { browser, element, by, Key } from 'protractor';
 
 describe('frontend App', () => {
